Add files via upload
Added ReturnToAddingItems use case (case a)- added boolean to local variables to indicate whether customer is currently adding items or not.  Included lines in payWithBanknote, payWithCoin and lookUpProductCode which sets this boolean to false, and returnToAddingItems changes it back to true. 

(Feel free to change this if needed for the GUI)